requirements.txt
fastapi
uvicorn
cryptography
pycryptodome
pyotp
requests